Update footer.html
Fix for the following error: "A file was included in `/_layouts/default.html` that is a symlink or does not exist in your `_includes` directory."

footer-custom.html was deleted earlier as it was used generated ads on site.
Its reference from the footer.html has to be deleted as well.